Added in navigation to Rewards tab

+++ app/src/main/java/com/example/android/turnip_habit_tracking_app/MainActivity.java
@@ -1,18 +1,16 @@
 package com.example.android.turnip_habit_tracking_app;
 
 import android.app.AlertDialog;
+import android.app.LoaderManager;
 import android.content.ContentValues;
+import android.content.CursorLoader;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.Loader;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
-import android.app.LoaderManager;
-import android.content.CursorLoader;
-import android.content.Loader;
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
 import android.support.v4.widget.CursorAdapter;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -51,15 +49,10 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
             });
             getLoaderManager().initLoader(0, null, this);
 
-            FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
+        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
 
 
-    }
 
-    private void insertHabit(String textHabit) {
-        ContentValues values = new ContentValues();
-        values.put(DBOpenHelper.HABIT_NAME, textHabit);
-        Uri habitURI = getContentResolver().insert(HabitsProvider.CONTENT_URI, values);
     }
 
     @Override
@@ -72,16 +65,26 @@ public class MainActivity extends AppCompatActivity implements LoaderManager.Loa
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
 
+
         int id = item.getItemId();
 
-        switch (id) {
-            case R.id.action_delete_all:
+        if  (id == R.id.action_delete_all) {
                 deleteAllHabits();
-                break;
         }
+
+        if (id == R.id.rewards){
+            openRewards();
+        }
+
 
         return super.onOptionsItemSelected(item);
     }
+
+    private void openRewards() {
+        Intent intent = new Intent (this, EditorActivity.class);
+        startActivity(intent);
+    }
+
 
     private void deleteAllHabits() {
         DialogInterface.OnClickListener dialogClickListener =
